backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.19.0
exceptiongroup==1.2.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pytest==6.2.5
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
referencing==0.35.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
toml==0.10.2
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.2
